+++ ignite/cmd/scaffold.go
@@ -96,7 +96,6 @@ with an "--ibc" flag. Note that the default module is not IBC-enabled.
 	c.AddCommand(NewScaffoldPacket())
 	c.AddCommand(NewScaffoldBandchain())
 	c.AddCommand(NewScaffoldVue())
-	c.AddCommand(NewScaffoldFlutter())
 	// c.AddCommand(NewScaffoldWasm())
 
 	return c