+++ changelog.md
@@ -6,6 +6,7 @@
 
 - [#2955](https://github.com/ignite/cli/pull/2955/) Add `ignite network request add-account` command.
 - [#2995](https://github.com/ignite/cli/pull/2995/) Add `ignite network request remove-validator` command.
+- [#2999](https://github.com/ignite/cli/pull/2999/) Add `ignite network request remove-account` command.
 
 ### Changes
 
@@ -14,6 +15,7 @@
 - [#2958](https://github.com/ignite/cli/pull/2958) Support absolute paths for client code generation config paths.
 - [#2993](https://github.com/ignite/cli/pull/2993) Hide `ignite scaffold band` command and deprecate functionality.
 - [#2986](https://github.com/ignite/cli/issues/2986) Remove `--proto-all-modules` flag from `serve` command.
+- [#2991](https://github.com/ignite/cli/pull/2991) Hide `ignite scaffold flutter` command and remove functionality.
 
 ## [`v0.25.1`](https://github.com/ignite/cli/releases/tag/v0.25.1)